+++ app.js
@@ -5,7 +5,7 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 var config = require('./config/config')
 
-var indexRouter = require('./routes/index');
+var indexRouter = require('./routes/api_router');
 
 var app = express();
