+++ common/cassandraFunc.js
