+++ controllers/v1/content_view.js
@@ -2,19 +2,22 @@ var config = require('../../config/config');
 var rp = require('request-promise');
 var request = require('request');
 var model = require('../../db')
-var helperFunc = require('../../helper/chart_data');
+var helperFunc = require('../../helper/chartData');
 
 //Controller for listing Top 5 contents viewed in platform
 exports.contentView = async function (req, res) {
     //get quey from cassandra
-    model.MyModel.findOneAsync({ qid: "content_viewed_in_platform_query" }, { allow_filtering: true })
+    model.MyModel.findOneAsync({ qid: "content_view_query" }, { allow_filtering: true })
         .then(async function (result) {
             var bodyParam = JSON.parse(result.query);
             if (config.druid.telemetry_datasource_name) {
                 bodyParam.dataSource = config.druid.telemetry_datasource_name;
             }
-            // get previous month date and append to intervals field
-            bodyParam.intervals = await getIntervals();
+            var today = new Date();
+            var mm = today.getMonth() + 1;
+            var year = today.getFullYear();
+            var date = year + "-" + mm + '-01T00:00:00.000Z';
+            bodyParam.filter.fields[1].lower = date;
             //Assign threshold value to restrict number of records to be shown
             bodyParam.threshold = config.druid.threshold_in_content_api;
             //pass the query as body param and get the result from druid
@@ -26,8 +29,7 @@ exports.contentView = async function (req, res) {
                 res.send({ "result": false, "data": [] })
             }
             else {
-                var responseObj = await helperFunc.contentViewResponeObj(data[0].result);
-                res.send(responseObj);
+                res.send({ "result": true, "data": data[0].result });
             }
         })
         .catch(function (err) {
@@ -45,7 +47,7 @@ exports.contentView = async function (req, res) {
 
 
 //Controller for listing Top 5 contents Downloaded by user in platform
-exports.contentDownloadedByUser = async function (req, res) {
+exports.contentViewedByUser = async function (req, res) {
     if (!req.body.usr_id) {
         res.status(400);
         var response = {
@@ -57,16 +59,21 @@ exports.contentDownloadedByUser = async function (req, res) {
     }
     else {
         //get quey from cassandra
-        model.MyModel.findOneAsync({ qid: "content_downloaded_by_user_query" }, { allow_filtering: true })
+        model.MyModel.findOneAsync({ qid: "content_viewed_by_user_query" }, { allow_filtering: true })
             .then(async function (result) {
                 var bodyParam = JSON.parse(result.query);
                 if (config.druid.telemetry_datasource_name) {
                     bodyParam.dataSource = config.druid.telemetry_datasource_name;
                 }
                 //append user id to the filter
-                 bodyParam.filter.fields[0].value = req.body.usr_id;
-                 // get previous month date and append to intervals field
-                 bodyParam.intervals = await getIntervals();
+                bodyParam.filter.fields[1].fields[1].fields[0].value = req.body.usr_id;
+                //calculate current date and month
+                var today = new Date();
+                var mm = today.getMonth() + 1;
+                var year = today.getFullYear();
+                var date = year + "-" + mm + '-01T00:00:00.000Z';
+                //append date to the filter
+                bodyParam.filter.fields[1].fields[0].lower = date;
                 //Assign threshold value to restrict number of records to be shown
                 bodyParam.threshold = config.druid.threshold_in_content_api;
                 //pass the query as body param and get the result from druid
@@ -78,11 +85,11 @@ exports.contentDownloadedByUser = async function (req, res) {
                     res.send({ "result": false, "data": [] })
                 }
                 else {
-                    var responseObj = await helperFunc.contentDownloadResponeObj(data[0].result);
-                    res.send(responseObj);
+                    res.send({ "result": true, "data": data[0].result });
                 }
             })
             .catch(function (err) {
+                console.log(err);
                 res.status(400);
                 var response = {
                     result: false,
@@ -106,11 +113,13 @@ exports.usageByContent = async function (req, res) {
             if (config.druid.telemetry_datasource_name) {
                 bodyParam.dataSource = config.druid.telemetry_datasource_name;
             }
-            
+            var today = new Date();
+            var mm = today.getMonth() + 1;
+            var year = today.getFullYear();
+            var date = year + "-" + mm + '-01T00:00:00.000Z';
+            bodyParam.filter.fields[0].value = date;
             //Assign threshold value to restrict number of records to be shown
             bodyParam.threshold = config.druid.threshold_in_content_api
-             // get previous month date and append to intervals field
-             bodyParam.intervals = await getIntervals();
             //pass the query as body param and get the result from druid
             var options = config.druid.options;
             options.method = "POST";
@@ -125,6 +134,7 @@ exports.usageByContent = async function (req, res) {
             }
         })
         .catch(function (err) {
+            console.log(err);
             res.status(400);
             var response = {
                 result: false,
@@ -135,24 +145,3 @@ exports.usageByContent = async function (req, res) {
         })
 
 }
-
-
-
-async function getIntervals() {
-    var now = new Date();
-    var prevMonthLastDate = new Date(now.getFullYear(), now.getMonth(), 0);
-    var prevMonthFirstDate = new Date(now.getFullYear() - (now.getMonth() > 0 ? 0 : 1), (now.getMonth() - 1 + 12) % 12, 1);
-
-    var formatDateComponent = function (dateComponent) {
-        return (dateComponent < 10 ? '0' : '') + dateComponent;
-    };
-
-    var formatDate = function (date) {
-        return date.getFullYear() + '-' + formatDateComponent(date.getMonth() + 1) + '-' + formatDateComponent(date.getDate()) + 'T00:00:00+00:00';
-    };
-
-   var intervals = formatDate(prevMonthFirstDate) + '/' + formatDate(prevMonthLastDate);
-
-   return intervals;
-
-}
