+++ controllers/v1/instance_observation.js
@@ -8,9 +8,9 @@ var client = new cassandra.Client({
   localDataCenter: 'datacenter1'
 });
 var model = require('../../db');
-var helperFunc = require('../../helper/chart_data');
-var commonCassandraFunc = require('../../common/cassandra_func');
-var pdfHandler = require('../../helper/common_handler');
+var helperFunc = require('../../helper/chartData');
+var commonCassandraFunc = require('../../common/cassandraFunc');
+var pdfHandler = require('../../helper/commonHandler');
 var ejs = require('ejs');
 var fs = require('fs');
 var uuidv4 = require('uuid/v4');
@@ -76,8 +76,6 @@ exports.instanceReport = async function (req, res) {
   }
 };
 
-
-// Function for instance observation PDF generation
 async function instancePdfFunc(req) {
   return new Promise(function (resolve, reject) {
     model.MyModel.findOneAsync({
@@ -86,6 +84,8 @@ async function instancePdfFunc(req) {
       allow_filtering: true
     })
       .then(async function (result) {
+
+        console.log("result", result);
 
         var bodyParam = JSON.parse(result.query);
         //bodyParam.dataSource = "sl_observation_dev";
@@ -131,6 +131,7 @@ exports.instancePdfReport = async function (req, res) {
     res.send(response);
   } else {
     reqData = req.query;
+    console.log("reqData",reqData)
     var dataReportIndexes = await commonCassandraFunc.checkReqInCassandra(reqData);
     // if(dataReportIndexes){
 
@@ -146,6 +147,9 @@ exports.instancePdfReport = async function (req, res) {
       dataReportIndexes.downloadpdfpath = dataReportIndexes.downloadpdfpath.replace(/^"(.*)"$/, '$1');
       let signedUlr = await pdfHandler.getSignedUrl(dataReportIndexes.downloadpdfpath);
 
+      // call to get SignedUrl 
+      console.log("instaRes=======", signedUlr);
+
       var response = {
         status: "success",
         message: 'Observation Pdf Generated successfully',
@@ -156,8 +160,8 @@ exports.instancePdfReport = async function (req, res) {
     } else {
       var instaRes = await instancePdfFunc(reqData);
 
-      if(("observationName" in instaRes) == true) { 
-      let resData = await pdfHandler.instanceObservationPdfGeneration(instaRes);
+      if(("observationName" in instaRes) == true) {      
+      let resData = await pdfHandler.pdfGeneration(instaRes);
 
       if (dataReportIndexes) {
         var reqOptions = {
