+++ views/header.ejs
@@ -4,7 +4,7 @@
             .observation-header {
                 text-align: center;
                 font-size: 12rem;
-                font-weight: 500;
+                font-weight: 600;
                 border-bottom: 1px solid #ccc;
                 width: 100%;
                 color: rgb(207, 207, 230);
