+++ views/multiSelectReport.ejs
@@ -1,5 +1,5 @@
 <div>
   
-    <img style="width:80%" src="<%= path %>" />
+    <img style="width:100%" src="<%= path %>" />
     <div style='page-break-after:always'></div>
 </div>
