+++ views/radioReport.ejs
@@ -1,4 +1,4 @@
 <div>
-    <img style="width:80%" src="<%= path %>" />
+    <img style="width:100%" src="<%= path %>" />
     <div style='page-break-after:always'></div>
 </div>
