+++ views/sliderReport.ejs
@@ -1,20 +1,20 @@
 <div class="container" style="padding:20px;margin:20px;">
     <% if(instaResponse.responseType == 'slider'){ %>
-    <div class="container" style="padding:20px;margin:20px">
-      <div style="border: solid 1px rgb(182, 175, 175)">
-        <h3 style="background-color: #eee;padding:10px; padding-left: 10px;font-size: 21px"><%= instaResponse.question %></h3>
-        <% for(var j=0; j < instaResponse.answers.length; j++) { %>
-          <div class="row" style="padding: 10px;padding-left:10px">
-              <div class="col-sm-12" style="font-size: 18px;">
-                <%= instaResponse.answers[j] %>
-          </div>
-        </div>
-      </div>
-    </div>
-    <% if(j == instaResponse.answers.length - 1){ %>
-    <p style="page-break-before: always;">&nbsp;</p>
+      <h3><%=  instaResponse.question %></h3>
+      <% for(var j=0; j < instaResponse.answers.length; j++) { %>
+            <div class="row">
+              <div class="col-sm-12">
+                <div class="col-sm-12" style="border: 1px solid #eae8e8;padding: 10px;">
+                  <div class="col-sm-12" style="font-size: 18px;">
+                    <%= instaResponse.answers[j] %>
+                  </div>
+                </div>
+              </div>
+            </div> 
+          <% if(j == instaResponse.answers.length - 1){ %>
+              <p style="page-break-before: always;">&nbsp;</p>
+          <% } %>
+      <% } %>
     <% } %>
-    <% } %>
-    <% } %>
-  
-  </div>
+    
+ </div>
